fix:adding react hook form to the calendar form and splitting form into its own component
