+++ src/components/forms/availabilities-form.tsx
@@ -1,0 +1,5 @@
+import React from "react";
+
+export default function AvailabilitiesForm() {
+  return <div>AvailabilitiesForm</div>;
+}
