+++ src/components/forms/bookings-form.tsx
@@ -1,0 +1,5 @@
+import React from "react";
+
+export default function BookingsForm() {
+  return <div>BookingsForm</div>;
+}
